clusterStatus move to checker
