+++ redis/redis-console/src/main/java/com/ctrip/xpipe/redis/console/migration/status/MigrationStatus.java
@@ -1,5 +1,6 @@
 package com.ctrip.xpipe.redis.console.migration.status;
 
+import com.ctrip.xpipe.redis.checker.migration.status.ClusterStatus;
 import com.ctrip.xpipe.redis.console.migration.model.MigrationCluster;
 import com.ctrip.xpipe.redis.console.migration.status.migration.*;
