+++ redis/redis-console/src/main/java/com/ctrip/xpipe/redis/console/service/impl/ClusterServiceWithoutDB.java
@@ -8,7 +8,7 @@ import com.ctrip.xpipe.redis.console.dto.ClusterDTO;
 import com.ctrip.xpipe.redis.console.dto.ClusterUpdateDTO;
 import com.ctrip.xpipe.redis.console.dto.MultiGroupClusterCreateDTO;
 import com.ctrip.xpipe.redis.console.dto.SingleGroupClusterCreateDTO;
-import com.ctrip.xpipe.redis.console.migration.status.ClusterStatus;
+import com.ctrip.xpipe.redis.checker.migration.status.ClusterStatus;
 import com.ctrip.xpipe.redis.console.model.*;
 import com.ctrip.xpipe.redis.console.model.consoleportal.ClusterDcGroupModel;
 import com.ctrip.xpipe.redis.console.model.consoleportal.ClusterListUnhealthyClusterModel;
