+++ redis/redis-console/src/test/java/com/ctrip/xpipe/redis/console/migration/model/DefaultMigrationClusterTest.java
@@ -5,7 +5,7 @@ import com.ctrip.xpipe.redis.console.migration.AbstractMigrationTest;
 import com.ctrip.xpipe.redis.console.migration.command.MigrationCommandBuilder;
 import com.ctrip.xpipe.redis.console.migration.model.impl.DefaultMigrationCluster;
 import com.ctrip.xpipe.redis.console.migration.model.impl.DefaultMigrationShard;
-import com.ctrip.xpipe.redis.console.migration.status.ClusterStatus;
+import com.ctrip.xpipe.redis.checker.migration.status.ClusterStatus;
 import com.ctrip.xpipe.redis.console.migration.status.MigrationStatus;
 import com.ctrip.xpipe.redis.console.migration.status.migration.MigrationCheckingState;
 import com.ctrip.xpipe.redis.console.migration.status.migration.MigrationPublishState;
