+++ redis/redis-console/src/test/java/com/ctrip/xpipe/redis/console/service/meta/impl/ClusterMetaServiceImplTest.java
@@ -4,7 +4,7 @@ import com.ctrip.xpipe.cluster.ClusterType;
 import com.ctrip.xpipe.cluster.DcGroupType;
 import com.ctrip.xpipe.redis.console.AbstractConsoleIntegrationTest;
 import com.ctrip.xpipe.redis.console.entity.AzGroupClusterEntity;
-import com.ctrip.xpipe.redis.console.migration.status.ClusterStatus;
+import com.ctrip.xpipe.redis.checker.migration.status.ClusterStatus;
 import com.ctrip.xpipe.redis.console.model.*;
 import com.ctrip.xpipe.redis.console.service.ReplDirectionService;
 import com.ctrip.xpipe.redis.console.service.ZoneService;
