+++ services/ctrip-integration-test/src/test/java/com/ctrip/xpipe/redis/ctrip/integratedtest/console/db/TransactionManagerTest.java
@@ -1,7 +1,7 @@
 package com.ctrip.xpipe.redis.ctrip.integratedtest.console.db;
 
 import com.ctrip.xpipe.redis.console.dao.MigrationEventDao;
-import com.ctrip.xpipe.redis.console.migration.status.ClusterStatus;
+import com.ctrip.xpipe.redis.checker.migration.status.ClusterStatus;
 import com.ctrip.xpipe.redis.console.model.ClusterTbl;
 import com.ctrip.xpipe.redis.console.service.ClusterService;
 import com.ctrip.xpipe.redis.console.service.migration.impl.MigrationRequest;
